+++ StarHubTH/StarHubTHApp.swift
@@ -1,7 +1,36 @@
 import SwiftUI
+import ApplicationServices
+
+class URLDispatcher: ObservableObject {
+    static let shared = URLDispatcher()
+    @Published var openedURL: URL? = nil
+}
+
+class AppDelegate: NSObject, NSApplicationDelegate {
+    func applicationDidFinishLaunching(_ notification: Notification) {
+        NSAppleEventManager.shared().setEventHandler(
+            self,
+            andSelector: #selector(handleGetURLEvent(event:withReplyEvent:)),
+            forEventClass: AEEventClass(kInternetEventClass),
+            andEventID: AEEventID(kAEGetURL)
+        )
+    }
+
+    @objc func handleGetURLEvent(event: NSAppleEventDescriptor, withReplyEvent: NSAppleEventDescriptor) {
+        if let urlString = event.paramDescriptor(forKeyword: AEKeyword(keyDirectObject))?.stringValue,
+           let url = URL(string: urlString) {
+            DispatchQueue.main.async {
+                URLDispatcher.shared.openedURL = url
+            }
+        }
+    }
+}
 
 @main
 struct StarHubTHApp: App {
+    @NSApplicationDelegateAdaptor(AppDelegate.self) var appDelegate
+    @StateObject private var urlDispatcher = URLDispatcher.shared
+    
     init() {
         if let currentLang = UserDefaults.standard.string(forKey: "currentLanguage") {
             UserDefaults.standard.set([currentLang], forKey: "AppleLanguages")
@@ -12,6 +41,7 @@ struct StarHubTHApp: App {
     var body: some Scene {
         WindowGroup {
             MainView()
+                .environmentObject(urlDispatcher)
         }
         .windowResizability(.contentMinSize)
         .defaultSize(width: 900, height: 600)